Fixed UI where mouse hover not registering
The UI buttons were not registering when the player would hover over a UI item in any of the menus. I still don't know what the issue was exactly but I replaced the Canvas and Event System and reconnected all the menu components to the manager/other pages. Still a lot more UI work to do to make it work entirely as intended.

The UI should set the selected object to null when mouse hover is detected. Although, if keyboard or gamepad movement is detected the cursor should go away and the first UI item should be selected.